teachlist finished
